# docs/swagger.yaml
basePath: /api
definitions:
  model.Artikel:
    properties:
      id:
        type: string
      id_kategori:
        type: string
      id_penulis:
        type: string
      isi:
        type: string
      judul:
        type: string
      tanggal:
        type: string
    type: object
  model.Kategori:
    properties:
      _id:
        description: 'contoh: "k001"'
        type: string
      deskripsi:
        type: string
      nama:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    email: kamu@email.com
    name: Developer API Support
  description: Dokumentasi REST API untuk manajemen artikel, kategori, komentar, dan
    penulis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Backend Artikel API
  version: "1.0"
paths:
...
  /api/artikels:
    get:
      consumes:
      - application/json
      description: Mengambil semua data artikel dari database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan semua artikel
      tags:
      - Artikel
    post:
      consumes:
      - application/json
      description: Menyimpan artikel baru ke database (butuh token)
      parameters:
      - description: Data artikel baru
        in: body
        name: artikel
        required: true
        schema:
          $ref: '#/definitions/model.Artikel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Membuat artikel baru
      tags:
      - Artikel
  /api/artikels/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus artikel tertentu dari database (butuh token)
      parameters:
      - description: ID Artikel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Menghapus artikel berdasarkan ID
      tags:
      - Artikel
    get:
      consumes:
      - application/json
      description: Mengambil artikel tertentu berdasarkan ID
      parameters:
      - description: ID Artikel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan artikel berdasarkan ID
      tags:
      - Artikel
    put:
      consumes:
      - application/json
      description: Mengubah data artikel berdasarkan ID (butuh token)
      parameters:
      - description: ID Artikel
        in: path
        name: id
        required: true
        type: string
      - description: Data artikel yang diupdate
        in: body
        name: artikel
        required: true
        schema:
          $ref: '#/definitions/model.Artikel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mengupdate artikel berdasarkan ID
      tags:
      - Artikel
  /api/kategoris:
    get:
      consumes:
      - application/json
      description: Mengambil semua data kategori dari database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan semua kategori
      tags:
      - Kategori
    post:
      consumes:
      - application/json
      description: Menambahkan kategori baru ke database (butuh token)
      parameters:
      - description: Data kategori baru
        in: body
        name: kategori
        required: true
        schema:
          $ref: '#/definitions/model.Kategori'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Membuat kategori baru
      tags:
      - Kategori
  /api/kategoris/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data kategori dari database berdasarkan ID (butuh token)
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Menghapus kategori berdasarkan ID
      tags:
      - Kategori
    get:
      consumes:
      - application/json
      description: Mengambil detail kategori berdasarkan ID
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan kategori berdasarkan ID
      tags:
      - Kategori
    put:
      consumes:
      - application/json
      description: Mengubah data kategori berdasarkan ID (butuh token)
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: string
      - description: Data kategori yang diupdate
        in: body
        name: kategori
        required: true
        schema:
          $ref: '#/definitions/model.Kategori'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mengupdate kategori berdasarkan ID
      tags:
      - Kategori
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"